Добавлен viewcontroller для перехода по ячейкам, настроен дефолтный вид контроллера по макету
